Move createElement resolution out of entry module
The implementation of createElement is different in production and
development. This moves the DEV check out of the entry module and into
the ReactElement module.

There was already a TODO comment for this; I'm doing it now because I'm
about to refactor the server entry module (ReactServer.js) to import
createElement directly instead of going through the client entry module
(React.js, which I'll also probably rename to ReactClient.js).
